# shared/api/openapi.yaml
openapi: 3.1.0
info:
  title: 112 AI Trainer REST API
  version: 0.3.0
  description: Backend scenario API and Voice Gateway API on separate ports.
servers:
- url: http://127.0.0.1:8000/api/v1
  description: Scenario backend
- url: http://127.0.0.1:8001/api/v1
  description: Voice Gateway
security:
- bearerAuth: []
paths:
...
  /instructor/groups/{gid}/insights:
    get:
      summary: Get Insights
      operationId: get_insights_api_v1_instructor_groups__gid__insights_get
      parameters:
      - name: gid
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Gid
      - name: X-User-Session
        in: header
        required: false
        schema:
          type: string
          default: ''
          title: X-User-Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/SavedInsight'
                - $ref: '#/components/schemas/EmptyInsight'
                title: Response Get Insights Api V1 Instructor Groups  Gid  Insights
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
    post:
      summary: Post Insights
      operationId: post_insights_api_v1_instructor_groups__gid__insights_post
      parameters:
      - name: gid
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Gid
      - name: X-User-Session
        in: header
        required: false
        schema:
          type: string
          default: ''
          title: X-User-Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedInsight'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    DifficultSkill:
      properties:
        label:
          type: string
          maxLength: 200
          minLength: 1
          title: Label
        explanation:
          type: string
          maxLength: 1000
          minLength: 1
          title: Explanation
        evidence_key:
          type: string
          maxLength: 80
          minLength: 1
          title: Evidence Key
      additionalProperties: false
      type: object
      required:
      - label
      - explanation
      - evidence_key
      title: DifficultSkill
    EmptyInsight:
      properties:
        status:
          type: string
          const: not_generated
          title: Status
        group_id:
          type: string
          title: Group Id
      type: object
      required:
      - status
      - group_id
      title: EmptyInsight
    InsightAggregate:
      properties:
        completed_attempts:
          type: integer
          minimum: 0.0
          title: Completed Attempts
        current_group_size:
          type: integer
          minimum: 0.0
          title: Current Group Size
        participant_count:
          anyOf:
          - type: integer
            minimum: 0.0
          - type: 'null'
          title: Participant Count
      type: object
      required:
      - completed_attempts
      - current_group_size
      title: InsightAggregate
    Recommendation:
      properties:
        title:
          type: string
          maxLength: 200
          minLength: 1
          title: Title
        rationale:
          type: string
          maxLength: 1000
          minLength: 1
          title: Rationale
        scenario_ids:
          items:
            type: string
          type: array
          maxItems: 5
          minItems: 1
          title: Scenario Ids
        evidence_keys:
          items:
            type: string
          type: array
          maxItems: 5
          minItems: 1
          title: Evidence Keys
      additionalProperties: false
      type: object
      required:
      - title
      - rationale
      - scenario_ids
      - evidence_keys
      title: Recommendation
    SavedInsight:
      properties:
        summary:
          type: string
          maxLength: 1500
          minLength: 1
          title: Summary
        difficult_skills:
          items:
            $ref: '#/components/schemas/DifficultSkill'
          type: array
          maxItems: 6
          title: Difficult Skills
        recommendations:
          items:
            $ref: '#/components/schemas/Recommendation'
          type: array
          maxItems: 6
          minItems: 1
          title: Recommendations
        limitations:
          items:
            type: string
          type: array
          maxItems: 5
          title: Limitations
        version:
          type: string
          const: group-insights-v1
          title: Version
        status:
          type: string
          enum:
          - ready
          - mock
          title: Status
        group_id:
          type: string
          title: Group Id
        created_at:
          type: string
          title: Created At
        provider:
          type: string
          title: Provider
        model:
          type: string
          title: Model
        data_fingerprint:
          type: string
          pattern: ^[a-f0-9]{64}$
          title: Data Fingerprint
        aggregate:
          $ref: '#/components/schemas/InsightAggregate'
        evidence:
          items:
            $ref: '#/components/schemas/InsightEvidence'
          type: array
          maxItems: 120
          title: Evidence
        stale:
          type: boolean
          title: Stale
        stale_reasons:
          items:
            type: string
          type: array
          title: Stale Reasons
      additionalProperties: false
      type: object
      required:
      - summary
      - recommendations
      - version
      - status
      - group_id
      - created_at
      - provider
      - model
      - data_fingerprint
      - aggregate
      - evidence
      - stale
      - stale_reasons
      title: SavedInsight
    InsightEvidence:
      properties:
        key:
          type: string
          title: Key
        kind:
          type: string
          enum:
          - field
          - sequence
          - scenario_performance
          title: Kind
        scenario_id:
          type: string
          title: Scenario Id
        label:
          anyOf:
          - type: string
          - type: 'null'
          title: Label
        configuration_revision:
          anyOf:
          - type: integer
          - type: 'null'
          title: Configuration Revision
        error_count:
          anyOf:
          - type: integer
          - type: 'null'
          title: Error Count
        eligible_attempts:
          anyOf:
          - type: integer
          - type: 'null'
          title: Eligible Attempts
        rate_percent:
          anyOf:
          - type: number
          - type: 'null'
          title: Rate Percent
        attempts:
          anyOf:
          - type: integer
          - type: 'null'
          title: Attempts
        average_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Average Score
      additionalProperties: false
      type: object
      required:
      - key
      - kind
      - scenario_id
      title: InsightEvidence
...